+++ core/transport/src/main/java/hera/client/TransactionEitherTemplate.java
@@ -23,6 +23,7 @@ import hera.api.tupleorerror.ResultOrError;
 import hera.api.tupleorerror.ResultOrErrorFuture;
 import hera.strategy.StrategyChain;
 import io.grpc.ManagedChannel;
+import java.math.BigInteger;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
@@ -57,7 +58,7 @@ public class TransactionEitherTemplate
           .apply(identify(transactionBaseTemplate.getCommitFunction(), TRANSACTION_COMMIT_EITHER));
 
   @Getter(lazy = true, value = AccessLevel.PROTECTED)
-  private final Function3<AccountAddress, AccountAddress, Long,
+  private final Function3<AccountAddress, AccountAddress, BigInteger,
       ResultOrErrorFuture<TxHash>> sendFunction =
           getStrategyChain()
               .apply(identify(transactionBaseTemplate.getSendFunction(), TRANSACTION_SEND_EITHER));
@@ -75,7 +76,7 @@ public class TransactionEitherTemplate
   @Override
   public ResultOrError<TxHash> send(final AccountAddress sender,
       final AccountAddress recipient,
-      final long amount) {
+      final BigInteger amount) {
     return getSendFunction().apply(sender, recipient, amount).get();
   }
